Move constants into each store module
Capital case action types

Also updated the action types to use capital casing.

Also updated components to use mapMutation vuex helper.

Feature: CMS-11254
