+++ src/MusicFestival.Vue.Template/Assets/Scripts/components/pages/LandingPage.vue
@@ -47,7 +47,8 @@ import LanguageSelector from '@/Scripts/components/widgets/LanguageSelector.vue'
 import EpiLink from '@/Scripts/components/widgets/EpiLink.vue';
 import Modal from '@/Scripts/components/widgets/Modal.vue';
 import BuyTicketBlock from '@/Scripts/components/blocks/BuyTicketBlock.vue';
-import { SHOW_MODAL } from '@/Scripts/store/mutation-types.js';
+import { SHOW_MODAL } from '@/Scripts/store/modules/appContext.js';
+import { mapMutations } from 'vuex';
 
 export default {
     components: {
@@ -59,11 +60,9 @@ export default {
         BuyTicketBlock
     },
     props: ['model'],
-    methods: {
-        showModal() {
-            this.$store.commit(SHOW_MODAL);
-        }
-    }
+    methods: mapMutations({
+        showModal: SHOW_MODAL
+    })
 };
 </script>
