--- src/MusicFestival.Vue.Template/Assets/Scripts/store/action-types.js
@@ -1,3 +1,0 @@
-// actions for the epiDataModel module
-export const updateModelByFriendlyUrl = 'updateModelByFriendlyUrl';
-export const updateModelByContentLink = 'updateModelByContentLink';
