--- src/MusicFestival.Vue.Template/Assets/Scripts/store/mutation-types.js
@@ -1,9 +1,0 @@
-// mutation for the epiDataModel module
-export const UPDATE_MODEL = 'UPDATE_MODEL';
-
-// mutation for the epiContext module
-export const UPDATE_CONTEXT = 'UPDATE_CONTEXT';
-
-//mutations for the appContext module
-export const SHOW_MODAL = 'SHOW_MODAL';
-export const HIDE_MODAL = 'HIDE_MODAL';
