+++ src/MusicFestival.Vue.Template/MusicFestival.Vue.Template.csproj
@@ -348,12 +348,10 @@
     <Content Include="Assets\Scripts\epiMessages.js" />
     <Content Include="Assets\Scripts\main.js" />
     <Content Include="Assets\Scripts\router.js" />
-    <Content Include="Assets\Scripts\store\action-types.js" />
     <Content Include="Assets\Scripts\store\index.js" />
     <Content Include="Assets\Scripts\store\modules\appContext.js" />
     <Content Include="Assets\Scripts\store\modules\epiContext.js" />
     <Content Include="Assets\Scripts\store\modules\epiDataModel.js" />
-    <Content Include="Assets\Scripts\store\mutation-types.js" />
     <Content Include="Content\editor.css" />
     <Content Include="postcss.config.js" />
     <Content Include="webpack.config.js" />
